Create NeedsFixing
Hey guys, maybe one of you can look at this and see what I'm missing. I got it working so that at least the messages send, they are encrypted so that if multiple messages are the same the ciphertext will always be different. The main problem is that I can't figure out why the ciphertext isn't decrypting between clients.